+++ ClientApp/src/app/main/tables/table-design/CreateEditTableColumnDialog.js
@@ -1,5 +1,5 @@
 import FuseLoading from "@fuse/core/FuseLoading";
-import { CircularProgress } from "@mui/material";
+import { CircularProgress, Icon, IconButton } from "@mui/material";
 import Button from "@mui/material/Button";
 import Dialog from "@mui/material/Dialog";
 import DialogActions from "@mui/material/DialogActions";
@@ -213,9 +213,22 @@ export const CreateEditTableColumnDialog = ({
             {...other}
         >
             <DialogTitle>
-                {defaultValues != null
-                    ? t("Update " + table)
-                    : t("Add " + table)}
+                <div style={{ display: "flex" }}>
+                    {defaultValues != null
+                        ? t("Update " + table)
+                        : t("Add " + table)}
+                    <div style={{ flex: "1 1 auto" }}></div>
+                    <IconButton
+                        className="min-w-auto"
+                        size="small"
+                        color="primary"
+                        onClick={() => {
+                            onClose();
+                        }}
+                    >
+                        <Icon>cancel</Icon>
+                    </IconButton>
+                </div>
             </DialogTitle>
             <DialogContent
                 dividers